feat: move pagetitle to common folder
