+++ src/app/components/input-copy-text/index.tsx
@@ -2,7 +2,7 @@ import { Col, Row } from 'antd';
 import { useTranslation } from 'react-i18next';
 
 import { InputCopyText } from '@/components/Atoms';
-import ComponentPageTitle from '@/components/Molecules/ComponentPageTitle';
+import ComponentPageTitle from '@/components/Common/ComponentPageTitle';
 
 export default function InputCopyTextPage() {
   const { t } = useTranslation();
