+++ src/app/components/select/index.tsx
@@ -3,7 +3,7 @@ import Row from 'antd/es/row';
 import { useTranslation } from 'react-i18next';
 
 import { Select } from '@/components/Atoms/Select';
-import ComponentPageTitle from '@/components/Molecules/ComponentPageTitle';
+import ComponentPageTitle from '@/components/Common/ComponentPageTitle';
 
 const OPTIONS = [
   {
